requirements-claim3.txt
python-dotenv==1.0.1
openai==1.43.0
